web_app/requirements.txt
catboost==1.0.3
imbalanced-learn==0.8.1
matplotlib==2.1.0
mysqlclient==2.0.3
nltk==3.2.5
Pillow==5.2.0
plotly==2.0.15
scikit-learn==0.24.2
joblib == 0.11
SQLAlchemy==1.2.19
sqlparse==0.3.0
tqdm==4.11.2
wordcloud==1.4.1
Flask==0.12.5
gunicorn==19.10.0
jsonschema==3.2.0
nbformat==5.1.3
numpy==1.16.0
pandas==1.1.5
plotly==2.0.15
python-dateutil==2.8.1